Create stats lib, and cleanup old excel/python name clash mappings
