+++ tests/fixtures/fixture.xlsx.yml
@@ -20,24 +20,24 @@ cell_map:
   Sheet1!A18: 18
   Sheet1!A19:
   Sheet1!A20:
-  Sheet1!B1: =xsum(_R_("Sheet1!A1:A3"))
-  Sheet1!B2: =xsum(_R_("Sheet1!A2:A4"))
-  Sheet1!B3: =xsum(_R_("Sheet1!A3:A5"))
-  Sheet1!B4: =xsum(_R_("Sheet1!A4:A6"))
-  Sheet1!B5: =xsum(_R_("Sheet1!A5:A7"))
-  Sheet1!B6: =xsum(_R_("Sheet1!A6:A8"))
-  Sheet1!B7: =xsum(_R_("Sheet1!A7:A9"))
-  Sheet1!B8: =xsum(_R_("Sheet1!A8:A10"))
-  Sheet1!B9: =xsum(_R_("Sheet1!A9:A11"))
-  Sheet1!B10: =xsum(_R_("Sheet1!A10:A12"))
-  Sheet1!B11: =xsum(_R_("Sheet1!A11:A13"))
-  Sheet1!B12: =xsum(_R_("Sheet1!A12:A14"))
-  Sheet1!B13: =xsum(_R_("Sheet1!A13:A15"))
-  Sheet1!B14: =xsum(_R_("Sheet1!A14:A16"))
-  Sheet1!B15: =xsum(_R_("Sheet1!A15:A17"))
-  Sheet1!B16: =xsum(_R_("Sheet1!A16:A18"))
-  Sheet1!B17: =xsum(_R_("Sheet1!A17:A19"))
-  Sheet1!B18: =xsum(_R_("Sheet1!A18:A20"))
+  Sheet1!B1: =sum_(_R_("Sheet1!A1:A3"))
+  Sheet1!B2: =sum_(_R_("Sheet1!A2:A4"))
+  Sheet1!B3: =sum_(_R_("Sheet1!A3:A5"))
+  Sheet1!B4: =sum_(_R_("Sheet1!A4:A6"))
+  Sheet1!B5: =sum_(_R_("Sheet1!A5:A7"))
+  Sheet1!B6: =sum_(_R_("Sheet1!A6:A8"))
+  Sheet1!B7: =sum_(_R_("Sheet1!A7:A9"))
+  Sheet1!B8: =sum_(_R_("Sheet1!A8:A10"))
+  Sheet1!B9: =sum_(_R_("Sheet1!A9:A11"))
+  Sheet1!B10: =sum_(_R_("Sheet1!A10:A12"))
+  Sheet1!B11: =sum_(_R_("Sheet1!A11:A13"))
+  Sheet1!B12: =sum_(_R_("Sheet1!A12:A14"))
+  Sheet1!B13: =sum_(_R_("Sheet1!A13:A15"))
+  Sheet1!B14: =sum_(_R_("Sheet1!A14:A16"))
+  Sheet1!B15: =sum_(_R_("Sheet1!A15:A17"))
+  Sheet1!B16: =sum_(_R_("Sheet1!A16:A18"))
+  Sheet1!B17: =sum_(_R_("Sheet1!A17:A19"))
+  Sheet1!B18: =sum_(_R_("Sheet1!A18:A20"))
   Sheet1!C1: =sin(_C_("Sheet1!B1") * (_C_("Sheet1!A1") ** 2))
   Sheet1!C2: =sin(_C_("Sheet1!B2") * (_C_("Sheet1!A2") ** 2))
   Sheet1!C3: =sin(_C_("Sheet1!B3") * (_C_("Sheet1!A3") ** 2))
@@ -57,3 +57,17 @@ cell_map:
   Sheet1!C17: =sin(_C_("Sheet1!B17") * (_C_("Sheet1!A17") ** 2))
   Sheet1!C18: =sin(_C_("Sheet1!B18") * (_C_("Sheet1!A18") ** 2))
   Sheet1!D1: =linest(_R_("Sheet1!C1:C18"), _R_("Sheet1!B1:B18"), degree=1)[0]
+  # Older mappings for excel functions that match Python built-in and keywords
+  Sheet2!A1: =x_abs(-1)
+  Sheet2!A2: =x_and(True)
+  Sheet2!A3: =xatan2(1, 1)
+  Sheet2!A4: =x_if(1, 2, 3)
+  Sheet2!A5: =x_int(1.23)
+  Sheet2!A6: =x_len("Plugh")
+  Sheet2!A7: =xmax(2, 3)
+  Sheet2!A8: =x_not(True)
+  Sheet2!A9: =x_or(True, False)
+  Sheet2!A10: =xmin(2, 3)
+  Sheet2!A11: =x_round(2.34)
+  Sheet2!A12: =xsum(1, 2)
+  Sheet2!A13: =x_xor(True, True)
