+++ tests/test_excelcompiler.py
@@ -8,6 +8,7 @@
 #   https://www.gnu.org/licenses/gpl-3.0.en.html
 
 import json
+import math
 import os
 import shutil
 from unittest import mock
@@ -744,6 +745,47 @@ def test_validate_count():
     assert excel_compiler.evaluate('Sheet!B1:B4') == (1, 2, 3, 3)
 
 
+def test_obsolete_function_mappings(fixture_dir):
+    """Verify that we are mapping the previous mapping to new function names
+
+    # Older mappings for excel functions that match Python built-in and keywords
+    old_map = {
+        "abs": "x_abs",
+        "and": "x_and",
+        "atan2": "xatan2",
+        "if": "x_if",
+        "int": "x_int",
+        "len": "x_len",
+        "max": "xmax",
+        "not": "x_not",
+        "or": "x_or",
+        "min": "xmin",
+        "round": "x_round",
+        "sum": "xsum",
+        "xor": "x_xor",
+    }
+    """
+    excel_compiler = ExcelCompiler.from_file(
+        os.path.join(fixture_dir, 'fixture.xlsx.yml'))
+
+    result = [excel_compiler.evaluate(f'Sheet2!A{i}') for i in range(1, 14)]
+    assert result == [
+        abs(-1),
+        True,
+        math.atan2(1, 1),
+        2,
+        int(1.23),
+        len("Plugh"),
+        max(2, 3),
+        not True,
+        True or False,
+        min(2, 3),
+        round(2.34),
+        sum((1, 2)),
+        True ^ True,
+    ]
+
+
 @pytest.mark.parametrize(
     'msg, formula', (
         ("Function XYZZY is not implemented. "
@@ -828,15 +870,15 @@ def test_plugins(excel_compiler):
             calc_and_check()
 
     with mock.patch('pycel.excelformula.ExcelFormula.default_modules', ()):
-        excel_compiler._plugin_modules = ('pycel.excellib', )
+        excel_compiler._plugin_modules = ('pycel.excellib', 'pycel.lib.stats')
         calc_and_check()
 
-    with mock.patch('pycel.excelformula.ExcelFormula.default_modules', ()):
-        excel_compiler._plugin_modules = 'pycel.excellib'
+    with mock.patch('pycel.excelformula.ExcelFormula.default_modules', ('pycel.excellib',)):
+        excel_compiler._plugin_modules = 'pycel.lib.stats'
         calc_and_check()
 
     with mock.patch('pycel.excelformula.ExcelFormula.default_modules',
-                    ('pycel.excellib', )):
+                    ('pycel.excellib', 'pycel.lib.stats')):
         excel_compiler._plugin_modules = None
         calc_and_check()
