+++ tests/test_excelformula.py
@@ -58,31 +58,31 @@ range_inputs = [
     FormulaTest(
         '=SUM(B5:B15)',
         'B5:B15|SUM',
-        'xsum(_R_("B5:B15"))'),
+        'sum_(_R_("B5:B15"))'),
     FormulaTest(
         '=SUM(B5:B15,D5:D15)',
         'B5:B15|D5:D15|SUM',
-        'xsum(_R_("B5:B15"), _R_("D5:D15"))'),
+        'sum_(_R_("B5:B15"), _R_("D5:D15"))'),
     FormulaTest(
         '=SUM(B5:B15 A7:D7)',
         'B5:B15|A7:D7| |SUM',
-        'xsum(_R_(str(_REF_("B5:B15") & _REF_("A7:D7"))))'),
+        'sum_(_R_(str(_REF_("B5:B15") & _REF_("A7:D7"))))'),
     FormulaTest(
         '=SUM((A:A,1:1))',
         'A:A|1:1|,|SUM',
-        'xsum(_R_("A:A"), _R_("1:1"))'),
+        'sum_(_R_("A:A"), _R_("1:1"))'),
     FormulaTest(
         '=SUM((A:A A1:B1))',
         'A:A|A1:B1| |SUM',
-        'xsum(_R_(str(_REF_("A:A") & _REF_("A1:B1"))))'),
+        'sum_(_R_(str(_REF_("A:A") & _REF_("A1:B1"))))'),
     FormulaTest(
         '=SUM(D9:D11,E9:E11,F9:F11)',
         'D9:D11|E9:E11|F9:F11|SUM',
-        'xsum(_R_("D9:D11"), _R_("E9:E11"), _R_("F9:F11"))'),
+        'sum_(_R_("D9:D11"), _R_("E9:E11"), _R_("F9:F11"))'),
     FormulaTest(
         '=SUM((D9:D11,(E9:E11,F9:F11)))',
         'D9:D11|E9:E11|F9:F11|,|,|SUM',
-        'xsum(_R_("D9:D11"), (_R_("E9:E11"), _R_("F9:F11")))'),
+        'sum_(_R_("D9:D11"), (_R_("E9:E11"), _R_("F9:F11")))'),
     FormulaTest(
         '=SUMIF(A1:A5, ">""", B1:B5)',
         'A1:A5|">"""|B1:B5|SUMIF',
@@ -90,7 +90,7 @@ range_inputs = [
     FormulaTest(
         '={SUM(B2:D2*B3:D3)}',
         'B2:D2|B3:D3|*|SUM|ARRAYROW|ARRAY',
-        '((xsum(_R_("B2:D2") * _R_("B3:D3")),),)'),
+        '((sum_(_R_("B2:D2") * _R_("B3:D3")),),)'),
     FormulaTest(
         '=RIGHT({"A","B"},A2:B2)',
         '"A"|"B"|ARRAYROW|ARRAY|A2:B2|RIGHT',
@@ -106,7 +106,7 @@ range_inputs = [
     FormulaTest(
         '=SUM(123 + SUM(456) + (45<6))+456+789',
         '123|456|SUM|+|45|6|<|+|SUM|456|+|789|+',
-        '(xsum((123 + xsum(456)) + (45 < 6)) + 456) + 789'),
+        '(sum_((123 + sum_(456)) + (45 < 6)) + 456) + 789'),
     FormulaTest(
         '=AVG(((((123 + 4 + AVG(A1:A2))))))',
         '123|4|+|A1:A2|AVG|+|AVG',
@@ -117,7 +117,7 @@ basic_inputs = [
     FormulaTest(
         '=SUM((A:A 1:1))',
         'A:A|1:1| |SUM',
-        'xsum(_R_(str(_REF_("A:A") & _REF_("1:1"))))'),
+        'sum_(_R_(str(_REF_("A:A") & _REF_("1:1"))))'),
     FormulaTest(
         '=A1',
         'A1',
@@ -133,7 +133,7 @@ basic_inputs = [
     FormulaTest(
         '=atan2(A1,B1)',
         'A1|B1|atan2',
-        'xatan2(_C_("A1"), _C_("B1"))'),
+        'atan2_(_C_("A1"), _C_("B1"))'),
     FormulaTest(
         '=5*log(sin()+2)',
         '5|sin|2|+|log|*',
@@ -185,15 +185,15 @@ basic_inputs = [
     FormulaTest(
         '=SUM(B5:B15,D5:D15)%',
         'B5:B15|D5:D15|SUM|%',
-        'xsum(_R_("B5:B15"), _R_("D5:D15")) / 100'),
+        'sum_(_R_("B5:B15"), _R_("D5:D15")) / 100'),
     FormulaTest(
         '=AND(G3, 1)',
         'G3|1|AND',
-        'x_and(_C_("G3"), 1)'),
+        'and_(_C_("G3"), 1)'),
     FormulaTest(
         '=OR(TRUE, TRUE(), FALSE, FALSE())',
         'TRUE|TRUE|FALSE|FALSE|OR',
-        'x_or(True, True, False, False)'),
+        'or_(True, True, False, False)'),
     FormulaTest(
         '=--4',
         '4|-|-',
@@ -237,15 +237,15 @@ if_inputs = [
         '   "  more ""test"" text"',
         '"a"|"a"|"b"|ARRAYROW|"c"|#N/A|ARRAYROW|1|-|TRUE|ARRAYROW|ARRAY|=|'
         '"yes"|"no"|IF|"  more ""test"" text"|&',
-        'x_if("a" == (("a", "b",), ("c", "#N/A",), (-1, True,),), "yes", "no")'
+        'if_("a" == (("a", "b",), ("c", "#N/A",), (-1, True,),), "yes", "no")'
         ' & "  more \\"test\\" text"'),
     FormulaTest(
         '=IF(R13C3>DATE(2002,1,6),0,IF(ISERROR(R[41]C[2]),0,IF(R13C3>=R[41]C[2]'
         ',0, IF(AND(R[23]C[11]>=55,R[24]C[11]>=20),R53C3,0))))',
         'R13C3|2002|1|6|DATE|>|0|R[41]C[2]|ISERROR|0|R13C3|R[41]C[2]|>=|0|'
         'R[23]C[11]|55|>=|R[24]C[11]|20|>=|AND|R53C3|0|IF|IF|IF|IF',
-        'x_if(_C_("C13") > date(2002, 1, 6), 0, x_if(iserror(_C_("C42")), 0, '
-        'x_if(_C_("C13") >= _C_("C42"), 0, x_if(x_and('
+        'if_(_C_("C13") > date(2002, 1, 6), 0, if_(iserror(_C_("C42")), 0, '
+        'if_(_C_("C13") >= _C_("C42"), 0, if_(and_('
         '_C_("L24") >= 55, _C_("L25") >= 20), _C_("C53"), 0))))'),
     FormulaTest(
         '=IF(R[39]C[11]>65,R[25]C[42],ROUND((R[11]C[11]*IF(OR(AND('
@@ -258,24 +258,24 @@ if_inputs = [
         'R[44]C[11]|R[43]C[11]|IF|*|R[14]C[11]|R[39]C[11]|55|>=|'
         'R[40]C[11]|20|>=|AND|R[40]C[11]|20|>=|R11C3|"YES"|=|AND|OR|'
         'R[45]C[11]|R[43]C[11]|IF|*|+|0|ROUND|IF',
-        'x_if(_C_("L40") > 65, _C_("AQ26"), x_round((_C_("L12") * x_if(x_or('
-        'x_and(_C_("L40") >= 55, _C_("L41") >= 20), x_and(_C_("L41") >= 20, '
+        'if_(_C_("L40") > 65, _C_("AQ26"), round_((_C_("L12") * if_(or_('
+        'and_(_C_("L40") >= 55, _C_("L41") >= 20), and_(_C_("L41") >= 20, '
         '_C_("C11") == "YES")), _C_("L45"), _C_("L44"))) + (_C_("L15") * '
-        'x_if(x_or(x_and(_C_("L40") >= 55, _C_("L41") >= 20), x_and(_C_("L41") '
+        'if_(or_(and_(_C_("L40") >= 55, _C_("L41") >= 20), and_(_C_("L41") '
         '>= 20, _C_("C11") == "YES")), _C_("L46"), _C_("L44"))), 0))'),
     FormulaTest(
         '=IF(AI119="","",E119)',
         'AI119|""|=|""|E119|IF',
-        'x_if(_C_("AI119") == "", "", _C_("E119"))'),
+        'if_(_C_("AI119") == "", "", _C_("E119"))'),
     FormulaTest(
         '=IF(P5=1.0,"NA",IF(P5=2.0,"A",IF(P5=3.0,"B",IF(P5=4.0,"C",'
         'IF(P5=5.0,"D",IF(P5=6.0,"E",IF(P5=7.0,"F",IF(P5=8.0,"G"))))))))',
         'P5|1.0|=|"NA"|P5|2.0|=|"A"|P5|3.0|=|"B"|P5|4.0|=|"C"|P5|5.0|=|'
         '"D"|P5|6.0|=|"E"|P5|7.0|=|"F"|P5|8.0|=|"G"|IF|IF|IF|IF|IF|IF|IF|IF',
-        'x_if(_C_("P5") == 1.0, "NA", x_if(_C_("P5") == 2.0, "A", '
-        'x_if(_C_("P5") == 3.0, "B", x_if(_C_("P5") == 4.0, "C", '
-        'x_if(_C_("P5") == 5.0, "D", x_if(_C_("P5") == 6.0, "E", '
-        'x_if(_C_("P5") == 7.0, "F", x_if(_C_("P5") == 8.0, "G"))))))))'),
+        'if_(_C_("P5") == 1.0, "NA", if_(_C_("P5") == 2.0, "A", '
+        'if_(_C_("P5") == 3.0, "B", if_(_C_("P5") == 4.0, "C", '
+        'if_(_C_("P5") == 5.0, "D", if_(_C_("P5") == 6.0, "E", '
+        'if_(_C_("P5") == 7.0, "F", if_(_C_("P5") == 8.0, "G"))))))))'),
 ]
 
 fancy_reference_inputs = [
@@ -303,7 +303,7 @@ fancy_reference_inputs = [
     FormulaTest(
         '=SUM(sheet1!$A$1:$B$2)',
         'sheet1!$A$1:$B$2|SUM',
-        'xsum(_R_("sheet1!A1:B2"))'),
+        'sum_(_R_("sheet1!A1:B2"))'),
     FormulaTest(
         '=[data.xls]sheet1!$A$1',
         '[data.xls]sheet1!$A$1',
@@ -330,7 +330,7 @@ fancy_reference_inputs = [
     FormulaTest(
         '=IF(configurations!$G$22=3,sizing!$C$303,M14)',
         'configurations!$G$22|3|=|sizing!$C$303|M14|IF',
-        'x_if(_C_("configurations!G22") == 3, _C_("sizing!C303"), _C_("M14"))'),
+        'if_(_C_("configurations!G22") == 3, _C_("sizing!C303"), _C_("M14"))'),
     FormulaTest(
         '=TableX[[#This Row],[COL1]]&"-"&TableX[[#This Row],[COL2]]',
         'TableX[[#This Row],[COL1]]|"-"|&|TableX[[#This Row],[COL2]]|&',
@@ -520,7 +520,7 @@ def test_multi_area_ranges(excel, ATestCell):
     with mock.patch.object(excel, '_defined_names', {
             'dname': (('$A$1', 's1'), ('$A$3:$A$4', 's2'))}):
         excel_formula = ExcelFormula('=sum(dname)', cell=cell)
-        assert excel_formula.ast.emit == 'xsum(_C_("s1!A1"), _R_("s2!A3:A4"))'
+        assert excel_formula.ast.emit == 'sum_(_C_("s1!A1"), _R_("s2!A3:A4"))'
 
 
 def test_str():
@@ -890,12 +890,12 @@ def test_row(formula, result, cell, empty_eval_context, ATestCell):
         ('=subtotal(01,A1:B3)', 'average(_R_("A1:B3"))'),
         ('=subtotal(02,A1:B3)', 'count(_R_("A1:B3"))'),
         ('=subtotal(03,A1:B3)', 'counta(_R_("A1:B3"))'),
-        ('=subtotal(04,A1:B3)', 'xmax(_R_("A1:B3"))'),
-        ('=subtotal(05,A1:B3)', 'xmin(_R_("A1:B3"))'),
+        ('=subtotal(04,A1:B3)', 'max_(_R_("A1:B3"))'),
+        ('=subtotal(05,A1:B3)', 'min_(_R_("A1:B3"))'),
         ('=subtotal(06,A1:B3)', 'product(_R_("A1:B3"))'),
         ('=subtotal(07,A1:B3)', 'stdev(_R_("A1:B3"))'),
         ('=subtotal(08,A1:B3)', 'stdevp(_R_("A1:B3"))'),
-        ('=subtotal(09,A1:B3)', 'xsum(_R_("A1:B3"))'),
+        ('=subtotal(09,A1:B3)', 'sum_(_R_("A1:B3"))'),
         ('=subtotal(10,A1:B3)', 'var(_R_("A1:B3"))'),
         ('=subtotal(11,A1:B3)', 'varp(_R_("A1:B3"))'),
     )
